// prueba3/angular-14-client/src/app/models/envio.model.ts
export class Envio {
    id?: any;
    clienteId?:any;
    productoId?:any;
    bodegaId?:any;
    puertoId?:any;
    cantidad?: number;
    precio?: number;
    descuento?: number;
    total?: number;
    fechaRegistro?: Date;
    fechaEntrega?: Date;
    placa?: string;
    flota?: string;
    guia?: string;
    internacional?: string;

}